# swiss: implement rehashInPlace