@petermattis 2024-02-10
Note that this PR builds on top of #3, so you'll want to review that one first. I think this data structure is at a good place after this commit. Performance remains much better than Go's builtin map and the worst case map growth scenarios from delete and put have been fixed with the implementation of `rehashInPlace`. 

Do either of you see anything else that remains to be done before trying to adopt usage of this map in Pebble and CRDB? Experimenting with SIMD seems like it can remain a TODO, IMO. I suppose we'll need to address the following TODO to use this map for the Pebble block cache:

> Abstract out the slice allocations so we can use manual memory allocation when used inside Pebble.

That doesn't feel difficult.